re-arranging transform api's under a dedicated file along with its docs
